Fixing the notification bug when there was no actions
And there is a known bug in Chrome that causes the notification badge to be overridden with the default bell icon when the app is installed. Fixed in Chrome 69, rollout scheduled on Sept 4. https://chromium.googlesource.com/chromium/src/+/21447c9d90951d0bded1a7a0c2ae29389d50b6eb